+++ module/CLI/test/Command/ShortUrl/GenerateShortUrlCommandTest.php
@@ -9,8 +9,10 @@ use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\UriInterface;
 use Shlinkio\Shlink\CLI\Command\ShortUrl\GenerateShortUrlCommand;
+use Shlinkio\Shlink\CLI\Util\ExitCodes;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Exception\InvalidUrlException;
+use Shlinkio\Shlink\Core\Exception\NonUniqueSlugException;
 use Shlinkio\Shlink\Core\Service\UrlShortener;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
@@ -35,7 +37,7 @@ class GenerateShortUrlCommandTest extends TestCase
     }
 
     /** @test */
-    public function properShortCodeIsCreatedIfLongUrlIsCorrect()
+    public function properShortCodeIsCreatedIfLongUrlIsCorrect(): void
     {
         $urlToShortCode = $this->urlShortener->urlToShortCode(Argument::cetera())->willReturn(
             (new ShortUrl(''))->setShortCode('abc123')
@@ -47,26 +49,41 @@ class GenerateShortUrlCommandTest extends TestCase
         ]);
         $output = $this->commandTester->getDisplay();
 
+        $this->assertEquals(ExitCodes::EXIT_SUCCESS, $this->commandTester->getStatusCode());
         $this->assertStringContainsString('http://foo.com/abc123', $output);
         $urlToShortCode->shouldHaveBeenCalledOnce();
     }
 
     /** @test */
-    public function exceptionWhileParsingLongUrlOutputsError()
+    public function exceptionWhileParsingLongUrlOutputsError(): void
     {
         $this->urlShortener->urlToShortCode(Argument::cetera())->willThrow(new InvalidUrlException())
                                                                ->shouldBeCalledOnce();
 
         $this->commandTester->execute(['longUrl' => 'http://domain.com/invalid']);
         $output = $this->commandTester->getDisplay();
-        $this->assertStringContainsString(
-            'Provided URL "http://domain.com/invalid" is invalid.',
-            $output
-        );
+
+        $this->assertEquals(ExitCodes::EXIT_FAILURE, $this->commandTester->getStatusCode());
+        $this->assertStringContainsString('Provided URL "http://domain.com/invalid" is invalid.', $output);
     }
 
     /** @test */
-    public function properlyProcessesProvidedTags()
+    public function providingNonUniqueSlugOutputsError(): void
+    {
+        $urlToShortCode = $this->urlShortener->urlToShortCode(Argument::cetera())->willThrow(
+            NonUniqueSlugException::class
+        );
+
+        $this->commandTester->execute(['longUrl' => 'http://domain.com/invalid', '--customSlug' => 'my-slug']);
+        $output = $this->commandTester->getDisplay();
+
+        $this->assertEquals(ExitCodes::EXIT_FAILURE, $this->commandTester->getStatusCode());
+        $this->assertStringContainsString('Provided slug "my-slug" is already in use', $output);
+        $urlToShortCode->shouldHaveBeenCalledOnce();
+    }
+
+    /** @test */
+    public function properlyProcessesProvidedTags(): void
     {
         $urlToShortCode = $this->urlShortener->urlToShortCode(
             Argument::type(UriInterface::class),
@@ -83,6 +100,7 @@ class GenerateShortUrlCommandTest extends TestCase
         ]);
         $output = $this->commandTester->getDisplay();
 
+        $this->assertEquals(ExitCodes::EXIT_SUCCESS, $this->commandTester->getStatusCode());
         $this->assertStringContainsString('http://foo.com/abc123', $output);
         $urlToShortCode->shouldHaveBeenCalledOnce();
     }